ai/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
ultralytics==8.4.31
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.3.0
python-multipart==0.0.9